refactor: move MergeAnswersForWorkflowResolution method to answer service and remove redundant workflow_merge.go file
